[mono-sybolicate] Use mono-symbolicate-native.dll
mono-symbolicate-tool depends on functions implemented natively.
Before this change we were assuming that mono-symbolicate would run only
with mono runtime and thus it was possible to use a runtime internal
method.

Nontheless we want to be able to run the tool on .NET too, so mono
symbolicate now depends on a native assembly called
mono-symbolicate-native.dll.

For simplicity the native assembly is just mono-sgen renamed.
Later on we can write build scripts for mono-symbolicate-native.dll that
will produce a much smaller assembly.